changed default permission level type from int to byte

+++ src/main/java/net/thenextlvl/tweaks/config/GeneralConfig.java
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName;
 
 public record GeneralConfig(
         @SerializedName("back-buffer-stack-size") int backBufferStackSize,
-        @SerializedName("default-permission-level") int defaultPermissionLevel,
+        @SerializedName("default-permission-level") byte defaultPermissionLevel,
         @SerializedName("override-join-message") boolean overrideJoinMessage,
         @SerializedName("override-quit-message") boolean overrideQuitMessage,
         @SerializedName("override-chat") boolean overrideChat,
